SA-1617: Adjusted loading state
